restructure shared code
